+++ CHANGELOG.md
@@ -13,6 +13,8 @@ SPDX-License-Identifier: AGPL-3.0-or-later
 - Simplify new/edit task experience. Extract the task title from the
   description.
 - Place task search in top to save space
+- You can now search tasks in all your workspace's projects
+  instead of having to look at each project individually.
 
 ### Breaking